requirements_pc.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
fire==0.4.0
Flask==2.1.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
opencv_python==4.5.1.48
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.3.25
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
compress_pickle==2.1.0
pycocotools==2.0.4
lmdb==1.3.0
pretrainedmodels
gdown
